requirements.txt
future~=0.18.2
opencv-python~=4.5.3.56
numpy~=1.21.0
pytesseract~=0.3.8
Pillow~=8.3.1
requests~=2.26.0
urllib3~=1.26.7
PyDirectInput~=1.0.4
pywin32~=301